+++ lib/directive/module.dart
@@ -26,6 +26,7 @@ part 'ng_style.dart';
 part 'ng_switch.dart';
 part 'ng_non_bindable.dart';
 part 'input_select.dart';
+part 'ng_control.dart';
 part 'ng_form.dart';
 part 'ng_model_validators.dart';
